requirements.txt
pygame==2.5.2
pyobjc-framework-Quartz==10.0
pytest==7.4.3